+++ priv/repo/migrations/20210604085943_add_locked_to_sites.exs
@@ -1,0 +1,9 @@
+defmodule Plausible.Repo.Migrations.AddLockedToSites do
+  use Ecto.Migration
+
+  def change do
+    alter table(:sites) do
+      add :locked, :boolean, null: false, default: false
+    end
+  end
+end